Move examples to separate directory
